Move common functionality to super trait for aggregates
