+++ datafusion/physical-expr/src/window/aggregate.rs
@@ -18,12 +18,10 @@
 //! Physical exec for aggregate window function expressions.
 
 use std::any::Any;
-use std::iter::IntoIterator;
 use std::ops::Range;
 use std::sync::Arc;
 
 use arrow::array::Array;
-use arrow::compute::{concat, SortOptions};
 use arrow::record_batch::RecordBatch;
 use arrow::{array::ArrayRef, datatypes::Field};
 
@@ -31,26 +29,23 @@ use datafusion_common::ScalarValue;
 use datafusion_common::{DataFusionError, Result};
 use datafusion_expr::{Accumulator, WindowFrame, WindowFrameUnits};
 
-use crate::window::window_expr::{reverse_order_bys, WindowFn, WindowFunctionState};
+use crate::window::window_expr::{reverse_order_bys, AggregateWindowExpr};
 use crate::window::{
     PartitionBatches, PartitionWindowAggStates, SlidingAggregateWindowExpr,
-    WindowAggState, WindowState,
 };
 use crate::{expressions::PhysicalSortExpr, PhysicalExpr};
 use crate::{window::WindowExpr, AggregateExpr};
 
-use super::window_frame_state::WindowFrameContext;
-
 /// A window expr that takes the form of an aggregate function
 #[derive(Debug)]
-pub struct AggregateWindowExpr {
+pub struct NonSlidingAggregateWindowExpr {
     aggregate: Arc<dyn AggregateExpr>,
     partition_by: Vec<Arc<dyn PhysicalExpr>>,
     order_by: Vec<PhysicalSortExpr>,
     window_frame: Arc<WindowFrame>,
 }
 
-impl AggregateWindowExpr {
+impl NonSlidingAggregateWindowExpr {
     /// create a new aggregate window function expression
     pub fn new(
         aggregate: Arc<dyn AggregateExpr>,
@@ -75,8 +70,7 @@ impl AggregateWindowExpr {
 /// peer based evaluation based on the fact that batch is pre-sorted given the sort columns
 /// and then per partition point we'll evaluate the peer group (e.g. SUM or MAX gives the same
 /// results for peers) and concatenate the results.
-
-impl WindowExpr for AggregateWindowExpr {
+impl WindowExpr for NonSlidingAggregateWindowExpr {
     /// Return a reference to Any that can be used for downcasting
     fn as_any(&self) -> &dyn Any {
         self
@@ -95,42 +89,7 @@ impl WindowExpr for AggregateWindowExpr {
     }
 
     fn evaluate(&self, batch: &RecordBatch) -> Result<ArrayRef> {
-        let sort_options: Vec<SortOptions> =
-            self.order_by.iter().map(|o| o.options).collect();
-        let mut row_wise_results: Vec<ScalarValue> = vec![];
-
-        let mut accumulator = self.aggregate.create_accumulator()?;
-        let length = batch.num_rows();
-        let (values, order_bys) = self.get_values_orderbys(batch)?;
-
-        let mut window_frame_ctx = WindowFrameContext::new(&self.window_frame);
-        let mut last_range = Range { start: 0, end: 0 };
-
-        // We iterate on each row to perform a running calculation.
-        // First, cur_range is calculated, then it is compared with last_range.
-        for i in 0..length {
-            let cur_range =
-                window_frame_ctx.calculate_range(&order_bys, &sort_options, length, i)?;
-            let value = if cur_range.end == cur_range.start {
-                // We produce None if the window is empty.
-                ScalarValue::try_from(self.aggregate.field()?.data_type())?
-            } else {
-                // Accumulate any new rows that have entered the window:
-                let update_bound = cur_range.end - last_range.end;
-                if update_bound > 0 {
-                    let update: Vec<ArrayRef> = values
-                        .iter()
-                        .map(|v| v.slice(last_range.end, update_bound))
-                        .collect();
-                    accumulator.update_batch(&update)?
-                }
-                accumulator.evaluate()?
-            };
-            row_wise_results.push(value);
-            last_range = cur_range;
-        }
-
-        ScalarValue::iter_to_array(row_wise_results.into_iter())
+        self.aggregate_evaluate(batch)
     }
 
     fn evaluate_stateful(
@@ -138,57 +97,17 @@ impl WindowExpr for AggregateWindowExpr {
         partition_batches: &PartitionBatches,
         window_agg_state: &mut PartitionWindowAggStates,
     ) -> Result<()> {
-        let field = self.aggregate.field()?;
-        let out_type = field.data_type();
-        for (partition_row, partition_batch_state) in partition_batches.iter() {
-            if !window_agg_state.contains_key(partition_row) {
-                let accumulator = self.aggregate.create_accumulator()?;
-                window_agg_state.insert(
-                    partition_row.clone(),
-                    WindowState {
-                        state: WindowAggState::new(
-                            out_type,
-                            WindowFunctionState::AggregateState(vec![]),
-                        )?,
-                        window_fn: WindowFn::Aggregate(accumulator),
-                    },
-                );
-            };
+        self.aggregate_evaluate_stateful(partition_batches, window_agg_state)?;
+        for partition_row in partition_batches.keys() {
             let window_state =
                 window_agg_state.get_mut(partition_row).ok_or_else(|| {
                     DataFusionError::Execution("Cannot find state".to_string())
                 })?;
-            let accumulator = match &mut window_state.window_fn {
-                WindowFn::Aggregate(accumulator) => accumulator,
-                _ => unreachable!(),
-            };
             let mut state = &mut window_state.state;
-            state.is_end = partition_batch_state.is_end;
-
-            let mut idx = state.last_calculated_index;
-            let mut last_range = state.window_frame_range.clone();
-            let mut window_frame_ctx = WindowFrameContext::new(&self.window_frame);
-            let out_col = self.get_result_column(
-                accumulator,
-                &partition_batch_state.record_batch,
-                &mut window_frame_ctx,
-                &mut last_range,
-                &mut idx,
-                state.is_end,
-            )?;
-            state.last_calculated_index = idx;
-            state.window_frame_range = last_range.clone();
-
-            state.out_col = concat(&[&state.out_col, &out_col])?;
-            let num_rows = partition_batch_state.record_batch.num_rows();
-            state.n_row_result_missing = num_rows - state.last_calculated_index;
-
-            state.window_function_state =
-                WindowFunctionState::AggregateState(accumulator.state()?);
-
-            // We can progress static start bounds also
-            // Since they are already calculated. Also we are sure that we
-            // will never call retract
+            // We can progress static start bounds, Since they are already calculated.
+            // Also we are sure that we will never call retract. They can be removed
+            // from state safely. (Enables us to run queries involving UNBOUNDED PRECEDING) to
+            // run with bounded memory
             if self.window_frame.start_bound.is_unbounded() {
                 state.window_frame_range.start = if state.window_frame_range.end >= 1 {
                     state.window_frame_range.end - 1
@@ -216,7 +135,7 @@ impl WindowExpr for AggregateWindowExpr {
         self.aggregate.reverse_expr().map(|reverse_expr| {
             let reverse_window_frame = self.window_frame.reverse();
             if reverse_window_frame.start_bound.is_unbounded() {
-                Arc::new(AggregateWindowExpr::new(
+                Arc::new(NonSlidingAggregateWindowExpr::new(
                     reverse_expr,
                     &self.partition_by.clone(),
                     &reverse_order_bys(&self.order_by),
@@ -236,13 +155,16 @@ impl WindowExpr for AggregateWindowExpr {
     fn uses_bounded_memory(&self) -> bool {
         // NOTE: Currently, groups queries do not support the bounded memory variant.
         self.aggregate.supports_bounded_execution()
-            // && !self.window_frame.start_bound.is_unbounded()
             && !self.window_frame.end_bound.is_unbounded()
             && !matches!(self.window_frame.units, WindowFrameUnits::Groups)
     }
 }
 
-impl AggregateWindowExpr {
+impl AggregateWindowExpr for NonSlidingAggregateWindowExpr {
+    fn get_accumulator(&self) -> Result<Box<dyn Accumulator>> {
+        self.aggregate.create_accumulator()
+    }
+
     /// For given range calculate accumulator result inside range on value_slice and
     /// update accumulator state
     fn get_aggregate_result_inside_range(
@@ -268,51 +190,5 @@ impl AggregateWindowExpr {
             accumulator.evaluate()?
         };
         Ok(value)
-    }
-
-    fn get_result_column(
-        &self,
-        accumulator: &mut Box<dyn Accumulator>,
-        record_batch: &RecordBatch,
-        window_frame_ctx: &mut WindowFrameContext,
-        last_range: &mut Range<usize>,
-        idx: &mut usize,
-        is_end: bool,
-    ) -> Result<ArrayRef> {
-        let (values, order_bys) = self.get_values_orderbys(record_batch)?;
-        // We iterate on each row to perform a running calculation.
-        let length = values[0].len();
-        let sort_options: Vec<SortOptions> =
-            self.order_by.iter().map(|o| o.options).collect();
-        let mut row_wise_results: Vec<ScalarValue> = vec![];
-        let field = self.aggregate.field()?;
-        let out_type = field.data_type();
-        while *idx < length {
-            let cur_range = window_frame_ctx.calculate_range(
-                &order_bys,
-                &sort_options,
-                length,
-                *idx,
-            )?;
-            // Exit if range end index is length, need kind of flag to stop
-            if cur_range.end == length && !is_end {
-                break;
-            }
-            let value = self.get_aggregate_result_inside_range(
-                last_range,
-                &cur_range,
-                &values,
-                accumulator,
-            )?;
-            row_wise_results.push(value);
-            last_range.start = cur_range.start;
-            last_range.end = cur_range.end;
-            *idx += 1;
-        }
-        Ok(if row_wise_results.is_empty() {
-            ScalarValue::try_from(out_type)?.to_array_of_size(0)
-        } else {
-            ScalarValue::iter_to_array(row_wise_results.into_iter())?
-        })
     }
 }
